+++ src/utils/encryptUtils.js
@@ -1,4 +1,4 @@
-import sodium from 'libsodium-wrappers'
+import sodium from 'libsodium-wrappers-sumo'
 import { Buffer } from 'buffer';